% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2mep.R
\name{buildDBids}
\alias{buildDBids}
\title{Creates a vector of collection labels for each gene module.}
\usage{
buildDBids(gmds)
}
\arguments{
\item{gmds}{A gene module database in the same format as created by
\code{importMSigDB.xml}.}
}
\value{
A vector of identifiers, one per gene module, with the
    format: "db_subdb".
}
\description{
Creates a vector of collection labels for each gene module.
}
\details{
This function is useful to subset a database of gene
    modules by collections.
}
\examples{
db <- readRDS(system.file("testgmd.RDS", package="gep2mep"))
ids <- buildDBids(db)

unique(ids)
## [1] "C3_TFT" "C3_MIR" "C4_CGN"

db <- db[ids=="C3_MIR"]

length(db)
## [1] 10

}
\seealso{
importMSigDB.xml
}
