% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2mep.R, R/gep2pep.R
\name{buildEmptyDB}
\alias{buildEmptyDB}
\alias{buildEmptyDB}
\title{Creates an empty DB of gene sets for converting GEPs.}
\usage{
buildEmptyDB(path, gmd, name = NULL, description = NULL)

buildEmptyDB(path, gmd, name = NULL, description = NULL)
}
\arguments{
\item{path}{Path to an empty folder where the repository will be
created.}

\item{gmd}{A list of gene modules (see \code{importMsigDB.xml}).}

\item{name}{Name of the repository. Defaults to \code{NULL} (a
generic name will be given).}

\item{description}{Description of the repository. Defaults to \code{NULL}
(a generic repository will be given).}

\item{path}{Path to an empty folder where the repository will be
created.}

\item{gmd}{A list of gene modules (see \code{importMsigDB.xml}).}

\item{name}{Name of the repository. Defaults to \code{NULL} (a
generic name will be given).}

\item{description}{Description of the repository. Defaults to \code{NULL}
(a generic repository will be given).}
}
\value{
An object of class \code{repo}.

An object of class \code{repo}.
}
\description{
Creates an empty DB of gene sets for converting GEPs.

Creates an empty DB of gene sets for converting GEPs.
}
\seealso{
buildPEPs

buildPEPs
}
