% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2pep.R
\name{buildPEPs}
\alias{buildPEPs}
\title{Build PEPs from GEPs and stores them in the repository.}
\usage{
buildPEPs(rp, geps, parallel = FALSE, existing = "stop")
}
\arguments{
\item{rp}{A repository created by \code{\link{createRepository}}.}

\item{geps}{A matrix of ranks where each row corresponds to a gene
and each column to a perturbagen. Each column must include all
ranks from 1 to the number of rows. Row and column names must
be defined. Row names must match names used in pathways
(unrecognized names will not be used).}

\item{parallel}{If TRUE, gene sets will be processed in
parallel. Requires a parallel backend.}

\item{existing}{What to do if PEPs for a given DB of gene sets are
    already present. Can be one of:

\itemize{

  \item{stop: }{the default, throws an error. This is the safest
  approach.}

  \item{skip: }{the existing DB will be skipped. This is especially
  useful to build a repository incrementally by repeatedly using
  the same call to \code{buildPEPs}}.

  \item{overwrite: }{all the existing PEPs for the DB will be replaced by
  new PEPs.}

  \item{append: }{the new PEPs will be appended to the existing
  ones. Useful to update the repository.}

}}
}
\value{
Nothing. The computed PEPs will be available in the
    repository.
}
\description{
Build PEPs from GEPs and stores them in the repository.
}
\examples{
db <- readRDS(system.file("testgmd.RDS", package="gep2pep"))
repo_path <- file.path(tempdir(), "gep2pepTemp")

rp <- createRepository(repo_path, db)
## Repo root created.
## Repo created.
## [15:45:06] Storing pathway data for DB: C3_TFT
## [15:45:06] Storing pathway data for DB: C3_MIR
## [15:45:06] Storing pathway data for DB: C4_CGN

rp
##         ID Dims     Size
## C3_TFT_gmd   10 18.16 kB
## C3_MIR_gmd   10 17.25 kB
## C4_CGN_gmd   10   6.9 kB

## Loading sample gene expression profiles
geps <- readRDS(system.file("testgep.RDS", package="gep2pep"))

geps[1:3,1:3]
##       (+)_chelidonine (+)_isoprenaline (+/_)_catechin
## AKT3               88              117            417
## MED6              357              410             34
## NR2E3             383              121            453

buildPEPs(rp, geps)

rp
##           ID Dims     Size
##   C3_TFT_gmd   10 18.16 kB
##   C3_MIR_gmd   10 17.25 kB
##   C4_CGN_gmd   10   6.9 kB
##       C3_TFT    2  1.07 kB
##       C3_MIR    2  1.07 kB
##       C4_CGN    2  1.04 kB
unlink(repo_path, TRUE)

}
\seealso{
buildPEPs
}
