% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2pep.R
\docType{class}
\name{CategorizedCollection-class}
\alias{CategorizedCollection-class}
\title{A class to contain categorized gene set collection}
\description{
This class is a simple generalization of the
\code{BroadCollection} function of \code{GSEABase} to store gene
sets having assigned categories and subcategories that can be
different from those of the MSigDB.
}
\section{Slots}{

\describe{
\item{\code{category}}{A character defining the main category that the gene
set belongs to.}

\item{\code{subCategory}}{A character defining the secondary category that
the gene set belongs to.}
}}

