% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2pep.R
\name{checkRepository}
\alias{checkRepository}
\title{Check an existyng repository for consistency}
\usage{
checkRepository(rp)
}
\arguments{
\item{rp}{A repository created by \code{\link{createRepository}}.}
}
\value{
Nothing.
}
\description{
Check both repository data consistency (see \code{repo_check} from
the \code{repo} package) and specific gep2pep data consistency.
}
\examples{
db <- readRDS(system.file("testgmd.RDS", package="gep2pep"))
repo_path <- file.path(tempdir(), "gep2pepTemp")

rp <- createRepository(repo_path, db)
checkRepository(rp)

unlink(repo_path, TRUE)
}
