% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2mep.R
\name{dbStats}
\alias{dbStats}
\title{Shows repository statistics}
\usage{
dbStats(path = "MEPDB")
}
\arguments{
\item{path}{Path to the root directory of an existing repository.}
}
\value{
Nothing, used for side effects.
}
\description{
Shows repository statistics
}
