% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2mep.R
\name{exportPertSEA}
\alias{exportPertSEA}
\title{Export PertSEA results to XLS format}
\usage{
exportPertSEA(rp, results, outname = "PertSEA.xls")
}
\arguments{
\item{rp}{A repository created by \code{buildEmptyDB}.}

\item{results}{The output of \code{PertSEA}.}

\item{outname}{Name of the XLS file to be created.}
}
\value{
Nothing.
}
\description{
Export PertSEA results to XLS format
}
