% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2mep.R
\name{exportSEA}
\alias{exportSEA}
\title{Export PertSEA or ModSEA results to XLS format}
\usage{
exportSEA(rp, results, outname = NULL)
}
\arguments{
\item{rp}{A repository created by \code{\link{createRepository}}.}

\item{results}{The output of \code{PertSEA} or \code{ModSEA}.}

\item{outname}{Name of the XLS file to be created.}
}
\value{
Nothing.
}
\description{
Export PertSEA or ModSEA results to XLS format
}
\examples{

db <- readRDS(system.file("testgmd.RDS", package="gep2mep"))
repo_path <- file.path(tempdir(), "gep2mepTemp")

rp <- createRepository(repo_path, db)
geps <- readRDS(system.file("testgep.RDS", package="gep2mep"))
buildMEPs(rp, geps)

pgset <- c("(+)_chelidonine", "(+/_)_catechin")
psea <- PertSEA(rp, pgset)

exportSEA(rp, psea)

unlink(repo_path, TRUE)

}
\seealso{
PertSEA, ModSEA
}
