% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2mep.R
\docType{package}
\name{gep2mep-package}
\alias{gep2mep-package}
\title{gep2mep: working with Module Expression Profiles}
\description{
gep2mep: working with Module Expression Profiles
}
\details{
TODO
}
\author{
Francesco Napolitano \email{franapoli@gmail.com}
}
