% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2mep.R
\name{gep2mep}
\alias{gep2mep}
\title{Converts Gene Expression Profiles (GEPs) to Module-EPs (MEPs).}
\usage{
gep2mep(geps, gmd, parallel = F)
}
\arguments{
\item{geps}{A matrix of ranks where each row corresponds to a gene
and each column to a perturbagen. Each column must include all
ranks from 1 to the number of rows. Row and column names must
be defined.}

\item{gmd}{A database of gene modules. See \code{importMsigDB.xml}
for the format.}

\item{parallel}{If TRUE, gene sets will be processed in
parallel. Requires a parallel backend.}
}
\value{
A list of two matrices, one for Enrichment Scores and one
    for p-values. Each entry (i,j) refers to gene set i and
    perturbagen j.
}
\description{
Converts Gene Expression Profiles (GEPs) to Module-EPs (MEPs).
}
\seealso{
buildMEPs
}
