% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2pep.R
\docType{package}
\name{gep2pep-package}
\alias{gep2pep-package}
\alias{gep2pep}
\title{gep2pep: creation and analysis of Pathway Expression Profiles}
\description{
gep2pep: creation and analysis of Pathway Expression Profiles
}
\details{
Pathway Expression Profiles (PEPs) are based on the expression of
pathways (defined as sets of genes) as opposed to individual
genes. \code{gep2pep} supports the convertion of gene expression
profiles (GEPs) to PEPs and performs enrichment analysis of
pathways or perturbagens. \code{gep2pep} creates a local repository
of gene sets, which can also be imported from the MSigDB
database. The local repository is in the \code{repo} format. When a
GEP is passed to \code{\link{buildPEPs}}, it refers to the stored
database of pathways to convert the GEP to a PEP and permanently
store the latter.

One type of analysis that can be performed on PEPs and that is
directly supported by \code{gep2pep} is the Drug-Set Enrichment
Analysis (DSEA, see reference below). It finds pathways that
are consistently dysregulated by a set of drugs, as opposed to a
background of other drugs. Of course PEPs may refer to
non-pharmacological perturbagens (genetic perturbations, disease
states, etc.) for analogous analyses. See \code{\link{PertSEA}}
function.

A complementary approach is that of finding perturbagens that
consistently dysregulate a set of pathways. This is the
pathway-based version of the Gene Set Enrichment Analysis
(GSEA). See \code{\link{PathSEA}}.

Naming conventions:

\itemize{

  \item{pathway: }{any set of gene identifiers (not necessarily
  representing a molecular pathway).}

  \item{pathway collection: }{a set of pathways.}

  \item{pathway database: }{a set of pathway collections, like the
  MSigDB.}

  \item{Pathway Expression Profile (PEP): }{a ranked list of
  pathways, as converted from a gene expression profile (GEP)
  according to a pathway collection.}

  \item{perturbagen }{any condition inducing a GEP and therefore a
  PEP.}

  \item{gep2pep repository: }{a pathway database and possibly a
  related database of PEPs as created by the \code{gep2pep}
  package. It is implemented in \code{repo} format.}

}
}
\references{
Napolitano F. et al, Drug-set enrichment analysis: a
    novel tool to investigate drug mode of action. Bioinformatics
    32, 235-241 (2016).
}
\author{
Francesco Napolitano \email{franapoli@gmail.com}
}
