% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2mep.R
\name{getDBlist}
\alias{getDBlist}
\title{Returns the names of the module databases in a repository.}
\usage{
getDBlist(rp)
}
\arguments{
\item{rp}{A repository created by \code{buildEmptyDB}.}
}
\value{
Vector of names.
}
\description{
Returns the names of the module databases in a repository.
}
\details{
The names are obtained from the repository entries that
    are tagged with "gmd" and removing the "_gmd" suffix.
}
