% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2pep.R
\name{loadSampleGEP}
\alias{loadSampleGEP}
\title{Loads sample Gene Expression Profiles}
\usage{
loadSampleGEP()
}
\value{
Sample gene expression data
}
\description{
Loads sample Gene Expression Profiles
}
\examples{

geps <- loadSampleGEP()
dim(geps)
## [1] 485   5
}
