% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2pep.R
\name{loadSamplePWS}
\alias{loadSamplePWS}
\title{Loads sample pathway collections}
\usage{
loadSamplePWS()
}
\value{
Sample pathway collections in \code{GeneSetCollection}
    format
}
\description{
Loads sample pathway collections
}
\examples{

geps <- loadSampleGEP()
geps
## GeneSetCollection
##   names: AAANWWTGC_UNKNOWN, AAAYRNCTG_UNKNOWN, ..., MORF_BUB3 (30 total)
##   unique identifiers: MEF2C, ATP1B1, ..., SLBP (5778 total)
##   types in collection:
##     geneIdType: SymbolIdentifier (1 total)
##     collectionType: BroadCollection (1 total)

}
