% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2mep.R
\name{ModSEA}
\alias{ModSEA}
\title{Performs Module Set Enrichment Analysis (MSEA)}
\usage{
ModSEA(rp_meps, modsets, bgsets = "all", collections = "all", details = T)
}
\arguments{
\item{rp_meps}{A repository created with
\code{\link{createRepository}}, and containing MEPs created with
\code{\link{buildMEPs}}.}

\item{modsets}{A list where each entry must be a vector of module
IDs and must be named like a module database. MSEA will be
performed for each database separately.}

\item{bgsets}{Another list like \code{modsets}, representing the
statistical background for each database. If set to "all" (the
default), all modules not in \code{modsets} will be used.}

\item{collections}{The set of collections to use. If set to "all",
all collections in the repository will be used.}

\item{details}{If TRUE (default) details will be reported for each
perturbagen in \code{pgset}.}
}
\value{
A list of 2, by names "ModSEA" and "details". The "ModSEA"
    entry is a 2-columns matrix including ESs and p-values for each
    collection and perturbagen. The "details" entry reports the
    rank of each module in \code{modsets} for each perturbagen.
}
\description{
Performs Module Set Enrichment Analysis (MSEA)
}
\details{
ModSEA is the analogous of Gene Set Enrichment Analysis
    (GSEA), but for gene modules instead of single genes.
}
\examples{
db <- readRDS(system.file("testgmd.RDS", package="gep2mep"))
repo_path <- file.path(tempdir(), "gep2mepTemp")

rp <- createRepository(repo_path, db)
geps <- readRDS(system.file("testgep.RDS", package="gep2mep"))
buildMEPs(rp, geps)

modsets <- list(C3_TFT = c("M11607", "M10817", "M16694"),
                C4_CGN = c("M19723", "M5038", "M13419", "M1094"))

msea <- ModSEA(rp, modsets)
## [15:35:29] Working on DB: C3_TFT
## [15:35:29] Common module sets removed from bgset.
## [15:35:29] Col-ranking DB...
## [15:35:29] Computing enrichments...
## [15:35:29] done.
## [15:35:29] Working on DB: C4_CGN
## [15:35:29] Common module sets removed from bgset.
## [15:35:29] Col-ranking DB...
## [15:35:29] Computing enrichments...
## [15:35:29] done.

msea$ModSEA$C3_TFT
##                         ES        PV
## (_)_mk_801       0.7142857 0.1666667
## (_)_atenolol     0.7142857 0.1666667
## (+)_isoprenaline 0.5714286 0.4000000
## (+/_)_catechin   0.5714286 0.4000000
## (+)_chelidonine  0.3809524 0.8500000

unlink(repo_path, TRUE)

}
