% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2pep.R
\name{PathSEA}
\alias{PathSEA}
\title{Performs Pathway Set Enrichment Analysis (PSEA)}
\usage{
PathSEA(rp_peps, modsets, bgsets = "all", collections = "all",
  details = T)
}
\arguments{
\item{rp_peps}{A repository created with
\code{\link{createRepository}}, and containing PEPs created with
\code{\link{buildPEPs}}.}

\item{modsets}{A list where each entry must be a vector of pathway
IDs and must be named like a pathway database. PSEA will be
performed for each database separately.}

\item{bgsets}{Another list like \code{modsets}, representing the
statistical background for each database. If set to "all" (the
default), all pathways not in \code{modsets} will be used.}

\item{collections}{The set of collections to use. If set to "all",
all collections in the repository will be used.}

\item{details}{If TRUE (default) details will be reported for each
perturbagen in \code{pgset}.}
}
\value{
A list of 2, by names "PathSEA" and "details". The "PathSEA"
    entry is a 2-columns matrix including ESs and p-values for each
    collection and perturbagen. The "details" entry reports the
    rank of each pathway in \code{modsets} for each perturbagen.
}
\description{
Performs Pathway Set Enrichment Analysis (PSEA)
}
\details{
PathSEA is the analogous of Gene Set Enrichment Analysis
    (GSEA), but for pathways instead of single genes.
}
\examples{
db <- readRDS(system.file("testgmd.RDS", package="gep2pep"))
repo_path <- file.path(tempdir(), "gep2pepTemp")

rp <- createRepository(repo_path, db)
geps <- readRDS(system.file("testgep.RDS", package="gep2pep"))
buildPEPs(rp, geps)

modsets <- list(C3_TFT = c("M11607", "M10817", "M16694"),
                C4_CGN = c("M19723", "M5038", "M13419", "M1094"))

psea <- PathSEA(rp, modsets)
## [15:35:29] Working on DB: C3_TFT
## [15:35:29] Common pathway sets removed from bgset.
## [15:35:29] Col-ranking DB...
## [15:35:29] Computing enrichments...
## [15:35:29] done.
## [15:35:29] Working on DB: C4_CGN
## [15:35:29] Common pathway sets removed from bgset.
## [15:35:29] Col-ranking DB...
## [15:35:29] Computing enrichments...
## [15:35:29] done.

psea$PathSEA$C3_TFT
##                         ES        PV
## (_)_mk_801       0.7142857 0.1666667
## (_)_atenolol     0.7142857 0.1666667
## (+)_isoprenaline 0.5714286 0.4000000
## (+/_)_catechin   0.5714286 0.4000000
## (+)_chelidonine  0.3809524 0.8500000

unlink(repo_path, TRUE)

}
