% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2mep.R
\name{PertSEA}
\alias{PertSEA}
\title{Performs Perturbagen Set Enrichment Analysis (PSEA)}
\usage{
PertSEA(rp_meps, pgset, bgset = "all", collections = "all",
  details = TRUE)
}
\arguments{
\item{rp_meps}{A repository created with
\code{\link{createRepository}}, and containing MEPs created with
\code{\link{buildMEPs}}.}

\item{pgset}{A vector of names of perturbagens. Corresponding MEPs
must exist in all the gene module collections currently in
\code{rp}.}

\item{bgset}{The background against which to compare
\code{pgset}. If set to \code{all} (default), all the remaining
MEPs will be used. Corresponding MEPs must exist in all the
gene module collections currently in \code{rp}.}

\item{collections}{The set of collections to use. If set to "all",
all collections in the repository will be used.}

\item{details}{If TRUE (default) details will be reported for each
perturbagen in \code{pgset}.}
}
\value{
A list of 2, by names "PertSEA" and "details". The
    "PertSEA" entry is a 2-columns matrix including ESs and
    p-values (see details) for each gene module database and
    perturbagen. The "details" entry reports the rank of each
    perturbagen in \code{pgset} for each gene module.
}
\description{
Performs Perturbagen Set Enrichment Analysis (PSEA)
}
\details{
Perturbagen Set Enrichment Analysis (PSEA) can be seen as
    a Gene-SEA performed over rows (as opposed to columns) of a
    profile matrix. For each gene module, all perturbagens are
    ranked and the p-value of a Kolmogorov-Smirnov test is computed
    for the ranks assigned to perturbagens in \code{pgset} against
    the ranks assigned to perturbagens in \code{bgset}. A positive
    (negative) Enrichment Score (ES) is also computed to indicate
    if each gene module is UP- (DOWN-) regulated by \code{pgset} as
    compared to \code{bgset}. See reference.
}
\examples{
db <- readRDS(system.file("testgmd.RDS", package="gep2mep"))
repo_path <- file.path(tempdir(), "gep2mepTemp")

rp <- createRepository(repo_path, db)
geps <- readRDS(system.file("testgep.RDS", package="gep2mep"))
buildMEPs(rp, geps)

pgset <- c("(+)_chelidonine", "(+/_)_catechin")
psea <- PertSEA(rp, pgset)
## [16:26:58] Common perturbagens removed from bgset.
## [16:26:58] Working on DB: C3_TFT
## [16:26:58] Row-ranking DB...
## [16:26:58] Computing enrichments...
## [16:26:58] done.
## [16:26:58] Working on DB: C3_MIR
## [16:26:58] Row-ranking DB...
## [16:26:58] Computing enrichments...
## [16:26:58] done.
## [16:26:58] Working on DB: C4_CGN
## [16:26:58] Row-ranking DB...
## [16:26:58] Computing enrichments...
## [16:26:58] done.

psea$PertSEA$C3_TFT
##                ES  PV
## M5067   1.0000000 0.2
## M2501   1.0000000 0.2
## M3128  -0.6666667 0.6
## M11607  0.6666667 0.6
## M10817 -0.6666667 0.6
## M16694 -0.6666667 0.6
## M14463  0.6666667 0.6
## M14686 -0.6666667 0.6
## M12599  0.5000000 0.9
## M4831   0.3333333 1.0

unlink(repo_path, TRUE)

}
\references{
Napolitano F. et al, Drug-set enrichment analysis: a
    novel tool to investigate drug mode of action. Bioinformatics
    32, 235-241 (2016).
}
