% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2mep.R
\name{buildDBids}
\alias{buildDBids}
\title{Creates unique identifiers for gene module databases.}
\usage{
buildDBids(gmds)
}
\arguments{
\item{gmds}{A gene module database in the same format created by
\code{importMsigDB.xml}.}
}
\value{
A vector of identifiers created as "DB_SUBDB".
}
\description{
Creates unique identifiers for gene module databases.
}
\seealso{
importMsigDB.xml
}
