% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2mep.R, R/gep2pep.R
\name{buildPEPs}
\alias{buildPEPs}
\alias{buildPEPs}
\title{Build PEPs using an existing repository and stores them in it.}
\usage{
buildPEPs(rp, geps, parallel = F, existing = "stop")

buildPEPs(rp, geps, parallel = F, existing = "stop")
}
\arguments{
\item{rp}{A gep2mep repository (see \code{buildEmptyDB}).}

\item{geps}{A matrix of ranks where each row corresponds to a gene
and each column to a perturbagen. Each column must include all
ranks from 1 to the number of rows. Row and column names must
be defined.}

\item{parallel}{If TRUE, gene sets will be processed in
parallel. Requires a parallel backend.}

\item{existing}{What to do if PEPs for a given DB of gene sets are
    already present. Can be one of:

+ stop: the default, throws an error. This is the safest approach.

+ skip: the existing DB will be skipped. This is especially useful
to build a repository incrementally by repeatedly using the same
call to \code{buildPEPs}.

+ overwrite: all the existing PEPs for the DB will be replaced by
new PEPs.

+ append: the new PEPs will be appended to the existing
ones. Useful to update the repository.}

\item{rp}{A gep2mep repository (see \code{buildEmptyDB}).}

\item{geps}{A matrix of ranks where each row corresponds to a gene
and each column to a perturbagen. Each column must include all
ranks from 1 to the number of rows. Row and column names must
be defined.}

\item{parallel}{If TRUE, gene sets will be processed in
parallel. Requires a parallel backend.}

\item{existing}{What to do if PEPs for a given DB of gene sets are
    already present. Can be one of:

+ stop: the default, throws an error. This is the safest approach.

+ skip: the existing DB will be skipped. This is especially useful
to build a repository incrementally by repeatedly using the same
call to \code{buildPEPs}.

+ overwrite: all the existing PEPs for the DB will be replaced by
new PEPs.

+ append: the new PEPs will be appended to the existing
ones. Useful to update the repository.}
}
\value{
Nothing. The computed PEPs will be available in the
    repository.

Nothing. The computed PEPs will be available in the
    repository.
}
\description{
Build PEPs using an existing repository and stores them in it.

Build PEPs using an existing repository and stores them in it.
}
\seealso{
buildPEPs

buildPEPs
}
