% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2pep.R
\name{buildPEPs}
\alias{buildPEPs}
\title{Build PEPs from GEPs and stores them in the repository.}
\usage{
buildPEPs(rp, geps, parallel = FALSE, collections = "all",
  replace_existing = FALSE, progress_bar = TRUE)
}
\arguments{
\item{rp}{A repository created by \code{\link{createRepository}}.}

\item{geps}{A matrix of ranks where each row corresponds to a gene
and each column to a condition. Each column must include all
ranks from 1 to the number of rows. Row and column names must
be defined. Row names will be matched against gene identifiers
in the pathways collections, and unrecognized gene names will
not be used.}

\item{parallel}{If TRUE, gene sets will be processed in
parallel. Requires a parallel backend.}

\item{collections}{A subset of the collection names returned by
\code{getCollections}. If set to "all" (default), all the
collections in \code{rp} will be used.}

\item{replace_existing}{What to do if PEPs, identified by column
names of \code{geps} are already present in the repository. If
set to TRUE, they will be replaced, otherwise they will be
skipped and only PEPs of new conditions will be computed and
added. Either ways, will throw a warning.}

\item{progress_bar}{If set to TRUE (default) will show a progress
bar updated after coversion of each column of \code{geps}.}
}
\value{
Nothing. The computed PEPs will be available in the
    repository.
}
\description{
Given a matrix of ranked lists of genes (GEPs) and a \code{gep2pep}
repository, converts GEPs to PEPs and stores the latter in the
repository.
}
\examples{
db <- readRDS(system.file("testgmd.RDS", package="gep2pep"))
repo_path <- file.path(tempdir(), "gep2pepTemp")

rp <- createRepository(repo_path, db)
## Repo root created.
## Repo created.
## [15:45:06] Storing pathway data for collection: C3_TFT
## [15:45:06] Storing pathway data for collection: C3_MIR
## [15:45:06] Storing pathway data for collection: C4_CGN

rp
##          ID   Dims     Size
## C3_TFT_sets   10   18.16 kB
## C3_MIR_sets   10   17.25 kB
## C4_CGN_sets   10     6.9 kB

## Loading sample gene expression profiles
geps <- readRDS(system.file("testgep.RDS", package="gep2pep"))

geps[1:3,1:3]
##       (+)_chelidonine (+)_isoprenaline (+/_)_catechin
## AKT3               88              117            417
## MED6              357              410             34
## NR2E3             383              121            453

buildPEPs(rp, geps)

rp
##           ID  Dims     Size
##   C3_TFT_sets   10 18.16 kB
##   C3_MIR_sets   10 17.25 kB
##   C4_CGN_sets   10   6.9 kB
##        C3_TFT    2  1.07 kB
##        C3_MIR    2  1.07 kB
##        C4_CGN    2  1.04 kB

unlink(repo_path, TRUE)

}
\seealso{
buildPEPs
}
