% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2pep.R
\name{CategorizedCollection}
\alias{CategorizedCollection}
\title{Constructor method for objects of class CategorizedCollection.}
\usage{
CategorizedCollection(category = "uncategorized",
  subCategory = "uncategorized", ...)
}
\arguments{
\item{category}{A character defining the main category that the gene
set belongs to.}

\item{subcategory}{A character defining the secondary category that
the gene set belongs to.}
}
\description{
See \code{CategorizedCollection-class}.
}
\examples{

library(GSEABase)
gs1 <- GeneSet(setName="set1", setIdentifier="101")
collectionType(gs1) <- CategorizedCollection()
}
