% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2pep.R
\name{CondSEA}
\alias{CondSEA}
\title{Performs Condition Set Enrichment Analysis}
\usage{
CondSEA(rp_peps, pgset, bgset = "all", collections = "all",
  details = TRUE)
}
\arguments{
\item{rp_peps}{A repository created with
\code{\link{createRepository}}, and containing PEPs created
with \code{\link{buildPEPs}}.}

\item{pgset}{A vector of names of conditions. Corresponding PEPs
must exist in all the pathway collections currently in
\code{rp}.}

\item{bgset}{The background against which to compare
\code{pgset}. If set to \code{all} (default), all the remaining
PEPs will be used. If provided, the corresponding PEPs must
exist in all the pathway collections currently in \code{rp}.}

\item{collections}{A subset of the collection names returned by
\code{getCollections}. If set to "all" (default), all the
collections in \code{rp} will be used.}

\item{details}{If TRUE (default) rank details will be reported for
each condition in \code{pgset}.}
}
\value{
A list of 2, by names "CondSEA" and "details". The
    "CondSEA" entry is a 2-columns matrix including ESs and
    p-values (see details) for each pathway database and
    condition. The "details" entry reports the rank of each
    condition in \code{pgset} for each pathway.
}
\description{
Condition Set Enrichment Analysis (CondSEA) can be seen as a
Gene-SEA performed over rows (as opposed to columns) of a matrix of
GEPs. It tells how much a pathway is consistently dysregulated
under a set of conditions (such as a set of drug treatments,
disease states, cell types, etc.) when compared to a statistical
background of other conditions.
}
\details{
For each pathway, all conditions are ranked by how much
    they dysregulate it (from the most UP-regulating to the most
    DOWN-regulating). Then, a Kolmogorov-Smirnov (KS) test is
    performed to compare the ranks assigned to conditions in
    \code{pgset} against the ranks assigned to conditions in
    \code{bgset}. A positive (negative) Enrichment Score (ES) of
    the KS test indicates whether each pathway is UP- (DOWN-)
    regulated by \code{pgset} as compared to \code{bgset}. A
    p-value is associated to the ES.

    When PEPs are obtained from drug-induced gene expression
    profiles, \code{PathSEA} can be used to perform Drug-Set
    Enrichment Analysis [1].
}
\examples{
db <- readRDS(system.file("testgmd.RDS", package="gep2pep"))
db <- as.CategorizedCollection(db)
repo_path <- file.path(tempdir(), "gep2pepTemp")

rp <- createRepository(repo_path, db)
geps <- readRDS(system.file("testgep.RDS", package="gep2pep"))
buildPEPs(rp, geps)

pgset <- c("(+)_chelidonine", "(+/_)_catechin")
psea <- CondSEA(rp, pgset)
## [16:26:58] Common conditions removed from bgset.
## [16:26:58] Working on collection: C3_TFT
## [16:26:58] Row-ranking collection...
## [16:26:58] Computing enrichments...
## [16:26:58] done.
## [16:26:58] Working on collection: C3_MIR
## [16:26:58] Row-ranking collection...
## [16:26:58] Computing enrichments...
## [16:26:58] done.
## [16:26:58] Working on collection: C4_CGN
## [16:26:58] Row-ranking collection...
## [16:26:58] Computing enrichments...
## [16:26:58] done.

psea$CondSEA$C3_TFT
##                ES  PV
## M5067   1.0000000 0.2
## M2501   1.0000000 0.2
## M3128  -0.6666667 0.6
## M11607  0.6666667 0.6
## M16694 -0.6666667 0.6
## M14463  0.6666667 0.6
## M14686 -0.6666667 0.6
## M12599  0.5000000 0.9
## M10817 -0.5000000 0.9
## M4831   0.3333333 1.0

unlink(repo_path, TRUE)

}
\references{
[1] Napolitano F. et al, Drug-set enrichment analysis: a
    novel tool to investigate drug mode of action. Bioinformatics
    32, 235-241 (2016).
}
