% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2pep.R
\name{createRepository}
\alias{createRepository}
\title{Creates a repository of pathway collections.}
\usage{
createRepository(path, gmd, name = NULL, description = NULL)
}
\arguments{
\item{path}{Path to a non-existing directory where the repository
will be created.}

\item{gmd}{A database of pathways, see details.}

\item{name}{Name of the repository. Defaults to \code{NULL} (a
generic name will be given).}

\item{description}{Description of the repository. Defaults to
\code{NULL} (a generic repository will be given).}
}
\value{
An object of class \code{repo}.
}
\description{
Creates a repository of pathway collections.
}
\details{
\code{gmd} must be in the same format as output by
    \code{\link{importMSigDB.xml}}. It is a list where each item
    includes the following fields:

\itemize{

  \item{id: }{a unique identifier of the gene set}

  \item{name: }{a descriptive name of the gene set}

  \item{db: }{an ID for the database this gene set belongs to (for
  example "GO")}

  \item{subdb: }{an ID for the sub-database this gene set belongs
  to (for example "BP")}

  \item{organism: }{name of the organism (for example
  "homo sapiens")}

  \item{desc: }{text description of the gene set (typically one
  short sentence)}

  \item{desc_full: }{a long, detailed description of the gene set}

  \item{set: }{genes in the set, as a vector of (typically) gene
  symbols}
}
}
\examples{

db <- readRDS(system.file("testgmd.RDS", package="gep2pep"))
repo_path <- file.path(tempdir(), "gep2pepTemp")

rp <- createRepository(repo_path, db)
## Repo root created.
## Repo created.
## [15:45:06] Storing pathway data for DB: C3_TFT
## [15:45:06] Storing pathway data for DB: C3_MIR
## [15:45:06] Storing pathway data for DB: C4_CGN

rp
##         ID Dims     Size
## C3_TFT_gmd   10 18.16 kB
## C3_MIR_gmd   10 17.25 kB
## C4_CGN_gmd   10   6.9 kB

unlink(repo_path, TRUE)
}
\seealso{
buildPEPs
}
