% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2mep.R, R/gep2pep.R
\name{dbStats}
\alias{dbStats}
\alias{dbStats}
\title{Shows repository statistics}
\usage{
dbStats(path = "PEPDB")

dbStats(path = "PEPDB")
}
\arguments{
\item{path}{Path to the root directory of an existing repository.}

\item{path}{Path to the root directory of an existing repository.}
}
\value{
Nothing, used for side effects.

Nothing, used for side effects.
}
\description{
Shows repository statistics

Shows repository statistics
}
