% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2pep.R
\name{dummyFunction}
\alias{dummyFunction}
\title{Dummy function for parameter inheritance}
\usage{
dummyFunction(rp, rp_peps, collections)
}
\arguments{
\item{rp}{A repository created by \code{\link{createRepository}}.}

\item{rp_peps}{A repository created with
\code{\link{createRepository}}, and containing PEPs created
with \code{\link{buildPEPs}}.}

\item{collections}{A subset of the collection names returned by
\code{getCollections}. If set to "all" (default), all the
collections in \code{rp} will be used.}
}
\value{
Nothing
}
\description{
Dummy function for parameter inheritance
}

