% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2mep.R
\name{dummyFunction}
\alias{dummyFunction}
\title{Dummy function for parameter inheritance}
\usage{
dummyFunction(rp, rp_meps, dbs, collections)
}
\arguments{
\item{rp}{A repository created by \code{\link{createRepository}}.}

\item{rp_meps}{A repository created with
\code{\link{createRepository}}, and containing MEPs created with
\code{\link{buildMEPs}}.}

\item{dbs}{A set of module collection names as returned by
\code{getCollections}. If set to "all" (default), all the collections
in \code{rp} will be used.}

\item{collections}{The set of collections to use. If set to "all",
all collections in the repository will be used.}
}
\value{
Nothing
}
\description{
Dummy function for parameter inheritance
}
