% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2mep.R
\docType{package}
\name{gep2mep-package}
\alias{gep2mep-package}
\alias{gep2mep}
\title{gep2mep: creation and analysis of Module Expression Profiles}
\description{
gep2mep: creation and analysis of Module Expression Profiles
}
\details{
Gene-Module expression profiles (MEPs) are based on the expression
of gene modules (sets of genes) as opposed to individual
genes. \code{gep2mep} supports the convertion of gene expression
profiles (GEPs) to MEPs and performs enrichment analysis of modules
or perturbagens. \code{gep2mep} creates a local repository of gene
modules, which can also be imported from the MSigDB database. The
local repository is in the \code{repo} format. When a GEP is passed
to \code{\link{buildMEPs}}, it refers to the stored database of
modules to convert the GEP to a MEP and permanently store the
latter. 

One type of analysis that can be performed on MEPs and that is
directly supported by \code{gep2mep} is the Drug-Set Enrichment
Analysis (DSEA, see reference below). It finds gene modules that
are consistently dysregulated by a set of drugs, as opposed to a
background of other drugs. Of course MEPs may refer to
non-pharmacological perturbagens (genetic perturbations, disease
states, etc.) for analogous analyses. See \code{\link{PertSEA}}
function.

A complementary approach is that of finding perturbagens that
consistently dysregulate a set of gene modules. This is the
module-based version of the Gene Set Enrichment Analysis
(GSEA). See \code{\link{ModSEA}}.

Naming conventions:

\itemize{

  \item{gene module: }{a set of gene identifiers.}

  \item{module collection: }{a set of gene modules.}

  \item{module database: }{a set of module collections, like the
  MSigDB.}

  \item{Module Expression Profile (MEP): }{a ranked list of
  modules, as converted from a gene expression profile (GEP)
  according to a module collection.}

  \item{perturbagen }{any condition related to a GEP and therefore
  a MEP.}

  \item{gep2mep repository: }{a module database and possibly a
  related database of MEPs as created by the \code{gep2mep}
  package. It is implemented in \code{repo} format.}

}
}
\references{
Napolitano F. et al, Drug-set enrichment analysis: a
    novel tool to investigate drug mode of action. Bioinformatics
    32, 235-241 (2016).
}
\author{
Francesco Napolitano \email{franapoli@gmail.com}
}
