\name{gep2pep-package}
\alias{gep2pep-package}
\alias{gep2pep}
\docType{package}
\title{
\packageTitle{gep2pep}
}
\description{
\packageDescription{gep2pep}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gep2pep}
\packageIndices{gep2pep}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{gep2pep}

Maintainer: \packageMaintainer{gep2pep}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
