% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2mep.R, R/gep2pep.R
\name{importMsigDB.xml}
\alias{importMsigDB.xml}
\alias{importMsigDB.xml}
\title{Imports gene set information from an MsigDB XML file.}
\usage{
importMsigDB.xml(fname)

importMsigDB.xml(fname)
}
\arguments{
\item{fname}{Path to an XML file downloaded from MsigDB}

\item{fname}{Path to an XML file downloaded from MsigDB}
}
\value{
A list of gene set entries (see details)

A list of gene set entries (see details)
}
\description{
Imports gene set information from an MsigDB XML file.

Imports gene set information from an MsigDB XML file.
}
\details{
The format required by gep2mep for a database of gene sets
    entries, provided by this function, is a list where each item
    includes the following fields:

+ setid: a unique identifier of the gene set

+ setname: a descriptive name of the gene set

+ db: an ID for the database this gene set belongs to (for
example "GO")

+ subdb: an ID for the sub-database this gene set belongs to (for
example "BP")

+ organism: name of the organism (for example "homo sapiens")

+ desc: text description of the gene set (typically one short
sentence)

+ desc_full: a long, detailed description of the gene set

+ set: genes in the set, as a vector of (typically) gene symbols

The format required by gep2mep for a database of gene sets
    entries, provided by this function, is a list where each item
    includes the following fields:

+ setid: a unique identifier of the gene set

+ setname: a descriptive name of the gene set

+ db: an ID for the database this gene set belongs to (for
example "GO")

+ subdb: an ID for the sub-database this gene set belongs to (for
example "BP")

+ organism: name of the organism (for example "homo sapiens")

+ desc: text description of the gene set (typically one short
sentence)

+ desc_full: a long, detailed description of the gene set

+ set: genes in the set, as a vector of (typically) gene symbols
}
