% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2pep.R
\name{importMSigDB.xml}
\alias{importMSigDB.xml}
\title{Imports pathways data from an MSigDB XML file.}
\usage{
importMSigDB.xml(fname)
}
\arguments{
\item{fname}{Path to an XML file downloaded from MSigDB.}
}
\value{
A list of pathway entries (see
    \code{\link{createRepository}}).
}
\description{
Creates a \code{gep2mep} repository of pathway data using the XML
distribution of the MSigDB (see references).
}
\examples{
\dontrun{

## To run this example, first obtain the MSigDB database in XML
## format (see
## http://software.broadinstitute.org/gsea/downloads.jsp). It is
## assumed that the database is locally available as the file
## "msigdb_v6.0.xml".

db <- importMSigDB.xml("msigdb_v6.0.xml")

## The database is now in an acceptable format to create a local
## repository using createRepository

length(db)
## [1] 18643

head(names(db))
## [1] "M3128"  "M11607" "M12599" "M5067"  "M10817" "M16694"

str(db[[1]], nchar.max=20)
## List of 8
##  $ id       : chr "M3128"
##  $ name     : chr "AAANWWTGC_UNKNOWN"
##  $ db       : chr "C3"
##  $ subdb    : chr "TFT"
##  $ organism : chr "Homo sapiens"
##  $ desc     : chr "Genes having at lea"| __truncated__
##  $ desc_full: chr "Comprehensive ident"| __truncated__
##  $ set      : chr [1:193] "MEF2C" "ATP1B1" "RORA" "CITED2" ...

}

## A small sample of the MSigDB as imported by importMSigDB.xml is
## included in gep2pep:

db_sample <- readRDS(system.file("testgmd.RDS", package="gep2pep"))

str(db_sample[[1]], nchar.max=20)
## List of 8
##  $ id       : chr "M3128"
##  $ name     : chr "AAANWWTGC_UNKNOWN"
##  $ db       : chr "C3"
##  $ subdb    : chr "TFT"
##  $ organism : chr "Homo sapiens"
##  $ desc     : chr "Genes having at lea"| __truncated__
##  $ desc_full: chr "Comprehensive ident"| __truncated__
##  $ set      : chr [1:193] "MEF2C" "ATP1B1" "RORA" "CITED2" ...
}
\references{
\url{http://software.broadinstitute.org/gsea/downloads.jsp}
}
