% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2pep.R
\name{loadSamplePWS}
\alias{loadSamplePWS}
\title{Loads sample pathway collections}
\usage{
loadSamplePWS()
}
\value{
Sample pathway collections in \code{GeneSetCollection}
    format
}
\description{
Loads sample pathway collections
}
\examples{

geps <- loadSampleGEP()

}
