% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gep2pep.R
\name{makeCollectionIDs}
\alias{makeCollectionIDs}
\title{Creates a vector of collection labels for each pathway.}
\usage{
makeCollectionIDs(gmds)
}
\arguments{
\item{gmds}{A pathway database in the same format as created by
\code{importMSigDB.xml}.}
}
\value{
A vector of identifiers, one per pathway, with the format:
    "db_subdb".
}
\description{
Creates a vector of collection labels for each pathway.
}
\details{
This function is useful to subset a database of pathway by
    collections.
}
\examples{
db <- readRDS(system.file("testgmd.RDS", package="gep2pep"))
ids <- makeCollectionIDs(db)

unique(ids)
## [1] "C3_TFT" "C3_MIR" "C4_CGN"

db <- db[ids=="C3_MIR"]

length(db)
## [1] 10

}
\seealso{
importMSigDB.xml
}
